+++ fiddles/angular/fiddle-0000-Template/tests/test.js
@@ -1,0 +1,10 @@
+
+TestCase("fiddle-0000-Template", {
+    setUp: function () {
+    },
+    tearDown: function () {
+    },
+    "test 1": function () {
+        return true;
+    }
+});